+++ src/cmd/graph_pwsh/Execute-Script.ps1
@@ -1,0 +1,15 @@
+[CmdletBinding()]
+
+Param (
+    [Parameter(Mandatory = $True, HelpMessage = "Powershell script URL")]
+    [string]$ScriptURL
+)
+
+Invoke-WebRequest -Uri $ScriptURL -OutFile ./script.ps1
+
+Write-Host "Executing the following script"
+Write-Host "=============================="
+cat ./script.ps1
+Write-Host "=============================="
+
+./script.ps1