+++ src/internal/connector/exchange/exchange_data_collection.go
@@ -82,7 +82,7 @@ type Collection struct {
 
 	// LocationPath contains the path with human-readable display names.
 	// IE: "/Inbox/Important" instead of "/abcdxyz123/algha=lgkhal=t"
-	locationPath path.Path
+	locationPath *path.Builder
 
 	state data.CollectionState
 
@@ -98,7 +98,8 @@ type Collection struct {
 // or notMoved (if they match).
 func NewCollection(
 	user string,
-	curr, prev, location path.Path,
+	curr, prev path.Path,
+	location *path.Builder,
 	category path.CategoryType,
 	items itemer,
 	statusUpdater support.StatusUpdater,
@@ -138,7 +139,7 @@ func (col *Collection) FullPath() path.Path {
 
 // LocationPath produces the Collection's full path, but with display names
 // instead of IDs in the folders.  Only populated for Calendars.
-func (col *Collection) LocationPath() path.Path {
+func (col *Collection) LocationPath() *path.Builder {
 	return col.locationPath
 }
 
@@ -251,11 +252,10 @@ func (col *Collection) streamItems(ctx context.Context, errs *fault.Bus) {
 			defer wg.Done()
 			defer func() { <-semaphoreCh }()
 
-			item, info, err := getItemWithRetries(
+			item, info, err := col.items.GetItem(
 				ctx,
 				user,
 				id,
-				col.items,
 				fault.New(true)) // temporary way to force a failFast error
 			if err != nil {
 				// Don't report errors for deleted items as there's no way for us to
@@ -279,7 +279,7 @@ func (col *Collection) streamItems(ctx context.Context, errs *fault.Bus) {
 			}
 
 			info.Size = int64(len(data))
-			info.ParentPath = col.locationPath.Folder(true)
+			info.ParentPath = col.locationPath.String()
 
 			col.data <- &Stream{
 				id:      id,
@@ -298,21 +298,6 @@ func (col *Collection) streamItems(ctx context.Context, errs *fault.Bus) {
 	}
 
 	wg.Wait()
-}
-
-// get an item while handling retry and backoff.
-func getItemWithRetries(
-	ctx context.Context,
-	userID, itemID string,
-	items itemer,
-	errs *fault.Bus,
-) (serialization.Parsable, *details.ExchangeInfo, error) {
-	item, info, err := items.GetItem(ctx, userID, itemID, errs)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	return item, info, nil
 }
 
 // terminatePopulateSequence is a utility function used to close a Collection's data channel